完成了ExIconv中的estring

+++ ExIconv/include/exiconv.h
@@ -2,21 +2,21 @@
 * @Author: sxf
 * @Date:   2015-11-04 12:07:35
 * @Last Modified by:   sxf
-* @Last Modified time: 2015-11-05 09:27:28
+* @Last Modified time: 2015-12-15 09:34:35
 */
 
 #include <stdint.h>
 #include <stddef.h>
 #include <stdio.h>
 
-typedef uint32_t echar_t;
+typedef uint16_t echar_t;
 struct FILE;
 
 
 /**
- * @brief 将UTF-32的字符串转换为utf8编码，便于输出
+ * @brief 将UTF-16的字符串转换为utf8编码，便于输出
  * 
- * @param data UTF-32格式的字符串数组
+ * @param data UTF-16格式的字符串数组
  * @param outsize 转换后的C字符串长度
  * 
  * @return 转换后的字符串，会自动新malloc空间，用过后由调用者释放
@@ -26,10 +26,10 @@ conv2utf8 (const echar_t* data, size_t* outsize);
 
 
 /**
- * @brief 将utf8编码转换为内部UTF-32格式
+ * @brief 将utf8编码转换为内部UTF-16格式
  * 
  * @param data C风格字符串数组
- * @param outsize 转换后的UTF-32字符串长度
+ * @param outsize 转换后的UTF-16字符串长度
  * 
  * @return 转换后的字符串，会自动新malloc空间，用过后由调用者释放
  */
@@ -40,9 +40,9 @@ utf8conv2echar (const char* data, size_t* outsize);
 
 
 /**
- * @brief 将UTF-32的字符串转换为utf8编码，便于输出
+ * @brief 将UTF-16的字符串转换为utf8编码，便于输出
  * 
- * @param data UTF-32格式的字符串数组
+ * @param data UTF-16格式的字符串数组
  * @param outsize 转换后的C字符串长度
  * 
  * @return 转换后的字符串，会自动新malloc空间，用过后由调用者释放
@@ -52,10 +52,10 @@ echar2code (const echar_t* data, size_t* outsize, const char* encode);
 
 
 /**
- * @brief 将指定编码转换为内部UTF-32格式
+ * @brief 将指定编码转换为内部UTF-16格式
  * 
  * @param data C风格字符串数组
- * @param outsize 转换后的UTF-32字符串长度
+ * @param outsize 转换后的UTF-16字符串长度
  * 
  * @return 转换后的字符串，会自动新malloc空间，用过后由调用者释放
  */
@@ -66,7 +66,7 @@ code2echar (const char* data, size_t* outsize, const char* encode);
 
 
 /**
- * @brief 自动读取文件下的文本内容，识别文本编码，并自动转换位为UTF-32的内部编码格式
+ * @brief 自动读取文件下的文本内容，识别文本编码，并自动转换位为UTF-16的内部编码格式
  * 
  * @param f 文件指针，读模式打开
  * @param outsize 转换后的字符串长度
@@ -78,7 +78,7 @@ autoreadfile (FILE* f, size_t* outsize);
 
 
 /**
- * @brief 自动识别文本格式，并转换为内部使用的UTF-32
+ * @brief 自动识别文本格式，并转换为内部使用的UTF-16
  * 
  * @param data 传入的字符串原始数据
  * @param outsize 转换后的字符串长度
